--- php-exercise5/src/control/controlAccount.php
@@ -1,7 +1,0 @@
-<?php
-
-if($_SERVER['REQUEST_METHOD'] == "POST" && isset($_POST["logout"])){
-    
-    header("Location: ../../index.php");
-    exit();
-}
