+++ php-exercise5/src/model/model.php
@@ -1,19 +1,6 @@
 <?php
-namespace model;
-class Accounts {
-    public $usernames = ["peter", "kirsch", "madrid"];
-    public $passwords = ["peter123", "kirsch123", "madrid123"];
-
-    public function getUsernames(){
-        return $this->usernames;
-    }
-
-    public function getPasswords(){
-        return $this->passwords;
-    }
-}
-
-class Validation {
+namespace src\Model;
+class Model {
     private $usernames;
     private $passwords;
     private $userInput;
